refactor(ui): split draft recovery and re-auth logic out of EditorPage
